Some changes
- Removing reverse speed - need rework
- Hide Details when Item Holder is destroy
- Changing Piercing Attribute to float
- Add todo JSON Leaderboard

+++ Assets/Scripts/ScriptableObject/ProjectileModifier.cs
@@ -12,7 +12,7 @@ public class ProjectileModifier : Modifier
     #region Stat attribute
     public float MovingSpeed = 0;
     public float Damage = 0;
-    public int Piercing = 0;
+    public float Piercing = 0;
     #endregion
 
     #region Customise attribute
